requirements.txt
streamlit
pillow
pandas
numpy